+++ tavern/_plugins/rest/response.py
@@ -10,14 +10,9 @@ except ImportError:
 from requests.status_codes import _codes
 
 from tavern.schemas.extensions import get_wrapped_response_function
-from tavern.util.dict_util import (
-    recurse_access_key,
-    deep_dict_merge,
-    check_is_simple_value,
-)
+from tavern.util.dict_util import recurse_access_key, deep_dict_merge
 from tavern.util.exceptions import TestFailError
 from tavern.util import exceptions
-from tavern.util.jmespath_util import check_jmespath_match
 from tavern.response.base import BaseResponse, indent_err_text
 
 logger = logging.getLogger(__name__)
@@ -310,39 +305,3 @@ class RestResponse(BaseResponse):
             logger.debug("Saved %s for '%s' from response", saved, key)
 
         return saved
-
-    def _validate_jmespath(self, body):
-        """
-        Valid all available jmespath queries for this response
-
-        Args:
-            body (dict, list): parsed body from response
-
-        Returns:
-            dict: values to save for future requests
-        """
-        saved_values = {}
-        for path_block in self.expected.get("jmespath", []):
-            try:
-                value = check_jmespath_match(
-                    body, path_block["query"], path_block.get("expected", None)
-                )
-            except exceptions.JMESError as e:
-                self._adderr("Error matching JMES in response", e=e)
-                continue
-
-            save_as = path_block.get("save_as")
-            if save_as:
-                try:
-                    check_is_simple_value(value, path_block["query"])
-                except exceptions.InvalidQueryResultTypeError as e:
-                    self._adderr(
-                        "Invalid query result type to save (was {})".format(
-                            type(value)
-                        ),
-                        e=e,
-                    )
-                else:
-                    saved_values[save_as] = value
-
-        return saved_values
